pyproject.toml
[tool.poetry]
name = "modelmorph"
version = "0.1.0"
description = "ModelMorph is a powerful library for integrating LLMs in chatbots and AI assistants."
authors = ["Esteban Salazar <estebansalazar498@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/Butis498/ModelMorph"


[tool.poetry.dependencies]
python = ">=3.8"
annotated-types = "0.7.0"
certifi = "2024.6.2"
charset-normalizer = "3.3.2"
distro = "1.9.0"
dnspython = "2.6.1"
h11 = "0.14.0"
httpcore = "1.0.6"
httpx = "0.27.2"
idna = "3.7"
jiter = "0.6.1"
loguru = "0.7.2"
openai = "1.52.2"
pydantic = "2.9.2"
pydantic_core = "2.23.4"
pymongo = "4.8.0"
python-dotenv = "1.0.1"
pytz = "2024.2"
requests = "2.32.3"
six = "1.16.0"
sniffio = "1.3.1"
tqdm = "4.66.6"
typing_extensions = "4.12.2"
tzdata = "2024.2"
urllib3 = "2.2.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
black = "^21.9b0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

